Fix race conditions with some votes
Currently, because of some changes in game code calling a vote after game
ended (during map voting screen) might not work (this happens with a
votes that stop matches and changes gametypes). Basically vote executes
just fine, but than when players vote for a map game switches to that
particular map (which might be not even available in that game mode) as
result map overwrites some settings.

This basically delays a vote if it was called when game ended and will
execute it when next map begings (which was an old behavior).